+++ packages/admin/components/layout/header/advanced-settings-toggle/index.tsx
@@ -1,13 +1,31 @@
 import { useSettings } from '@/admin/contexts/settings-context';
 import { ToggleControl } from '@wordpress/components';
+import { useEffect, useState } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 
 /**
  * The advanced settings toggle.
  */
 export const AdvancedSettingsToggle = () => {
-	const { showAdvancedSettings, updateSettings, saveSettings } =
-		useSettings();
+	const {
+		showAdvancedSettings,
+		updateSettings,
+		saveSettings,
+		isSaving,
+		settings,
+	} = useSettings();
+	const [ needsSave, setNeedsSave ] = useState( false );
+
+	useEffect( () => {
+		const save = async () => {
+			if ( needsSave && ! isSaving ) {
+				await saveSettings( 'wpgraphql_login_settings' );
+				setNeedsSave( false );
+			}
+		};
+
+		save();
+	}, [ needsSave, isSaving, saveSettings ] );
 
 	const classNames = 'wp-graphql-headless-login__advanced-settings-toggle';
 	const label = __( 'Show advanced settings', 'wp-graphql-headless-login' );
@@ -16,12 +34,12 @@ export const AdvancedSettingsToggle = () => {
 		await updateSettings( {
 			slug: 'wpgraphql_login_settings',
 			values: {
+				...settings?.wpgraphql_login_settings,
 				show_advanced_settings: value,
 			},
 		} );
 
-		// Save the setting immediately
-		await saveSettings( 'wpgraphql_login_settings' );
+		setNeedsSave( true );
 	};
 
 	return (
@@ -29,7 +47,8 @@ export const AdvancedSettingsToggle = () => {
 			className={ classNames }
 			checked={ showAdvancedSettings }
 			label={ label }
-			onChange={ ( value ) => setShowAdvancedSettings( value ) }
+			disabled={ isSaving }
+			onChange={ setShowAdvancedSettings }
 		/>
 	);
 };