+++ CHANGELOG.md
@@ -5,10 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
-- ci: Test compatibility with WordPress 6.7.1.
 - feat!: Refactor Settings handling and storage.
-- feat: Refactor admin package.
 - feat: Add Upgrader class for handling plugin upgrades.
+- feat: Refactor admin package.
 - chore!: Bump minimum WordPress version to 6.2.
 - chore!: Remove `vendor`, `vendor-prefixed`, and `build` directories from the repository.
 - chore: Update Composer dependencies.
@@ -16,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - chore: Update Strauss to v0.19.4.
 - ci!: Update ci scripts and commands, docker configuration, and env variables.
 - tests!: Update codeception tests to use `wp-browser` > v3.5.x.
+- ci: Test compatibility with WordPress 6.7.1.
 - tests: Format and lint test files.
 
 ## [0.3.1] - 2024-09-07